Set up testing
I could not use a test framework because I want the tests to be run in
CLI (so with Node) *and* use ES6 modules without a transpiler.
Currently, Node supports ES6 modules for packages having the "module"
type and if it is run with the --experimental-modules flag.
Tests frameworks were not runnable in this context, so I wrote a basic
test runner relying on the native node assertion library.